+++ src/pgstac/migrations/pgstac.0.6.13-0.7.0.sql
@@ -116,8 +116,22 @@ SET ROLE pgstac_admin;
 
 SET SEARCH_PATH TO pgstac, public;
 
-DROP FUNCTION IF EXISTS analyze_items;
-DROP FUNCTION IF EXISTS validate_constraints;
+
+DO $$
+  BEGIN
+    DROP FUNCTION IF EXISTS analyze_items;
+  EXCEPTION WHEN others THEN
+    RAISE NOTICE '%, skipping', SQLERRM USING ERRCODE = SQLSTATE;
+  END
+$$;
+DO $$
+  BEGIN
+    DROP FUNCTION IF EXISTS validate_constraints;
+  EXCEPTION WHEN others THEN
+    RAISE NOTICE '%, skipping', SQLERRM USING ERRCODE = SQLSTATE;
+  END
+$$;
+
 SET client_min_messages TO WARNING;
 SET SEARCH_PATH to pgstac, public;
 -- BEGIN migra calculated SQL