+++ src/pgstac/sql/999_version.sql
@@ -1,1 +1,1 @@
-SELECT set_version('0.7.1');
+SELECT set_version('0.7.2');